added -r command line option

+++ src/importexport/internal/pngwriter.cpp
@@ -55,7 +55,7 @@ mu::Ret PngWriter::write(const Score& score, IODevice& destinationDevice, const 
         pageRect = page->tbbox() + margins;
     }
 
-    const double CANVAS_DPI = configuration()->exportPngDpiResolution();
+    const float CANVAS_DPI = configuration()->exportPngDpiResolution();
     int width = std::lrint(pageRect.width() * CANVAS_DPI / DPI);
     int height = std::lrint(pageRect.height() * CANVAS_DPI / DPI);
